+++ src/brasil/gov/agenda/config.py
@@ -17,6 +17,7 @@ class HiddenProducts(object):
             'brasil.gov.agenda.upgrades.v3000',
             'brasil.gov.agenda.upgrades.v4000',
             'brasil.gov.agenda.upgrades.v4001'
+            'brasil.gov.agenda.upgrades.v4002'
         ]