+++ src/brasil/gov/agenda/content/agendadiaria.py
@@ -1,22 +1,18 @@
 # -*- coding: utf-8 -*-
-
 from brasil.gov.agenda import _
 from brasil.gov.agenda import utils
 from brasil.gov.agenda.config import AGENDADIARIAFMT
-from brasil.gov.agenda.interfaces import IAgenda
 from brasil.gov.agenda.interfaces import IAgendaDiaria
 from brasil.gov.agenda.interfaces import ICompromisso
 from DateTime import DateTime
 from five import grok
-from plone.app.dexterity.behaviors.exclfromnav import IExcludeFromNavigation
 from plone.dexterity.content import Container
-from plone.directives import form
 from plone.indexer.decorator import indexer
 from plone.supermodel.interfaces import IDefaultFactory
-from zope.interface import provider
-from zope.schema.interfaces import IContextAwareDefaultFactory
 from z3c.form.validator import SimpleFieldValidator
 from zope.interface import Invalid
+from zope.interface import provider
+from zope.schema.interfaces import IContextAwareDefaultFactory
 
 
 class AgendaDiaria(Container):
@@ -33,13 +29,22 @@ class AgendaDiaria(Container):
         title = u'Agenda de %s para %s' % (autoridade, fmt_date)
         return title
 
+    def exclude_from_nav(self):
+        """ AgendaDiaria nao eh visivel na navegacao do portal
+        """
+        return True
 
-@form.default_value(field=IExcludeFromNavigation['exclude_from_nav'])
-def exclude_from_nav_default_value(data):
-    # AgendaDiaria e Compromisso nao devem aparecer na navegacao
-    context = data.context
-    exclude = IAgenda.providedBy(context) or IAgendaDiaria.providedBy(context)
-    return exclude
+    @property
+    def effective_date(self):
+        """ Metodo que retorna
+        """
+        # Usamos a logica do indexer para start_date
+        date = _start_date(self)
+        if not date.isPast():
+            # Se estiver no futuro, data de efetivacao sera agora
+            now = DateTime()
+            return now
+        return date
 
 
 @provider(IContextAwareDefaultFactory)
@@ -53,6 +58,11 @@ def default_autoridade(context):
 @provider(IContextAwareDefaultFactory)
 def default_location(context):
     return getattr(context, 'location', u'')
+
+
+@provider(IContextAwareDefaultFactory)
+def default_subjects(context):
+    return getattr(context, 'subjects', ())
 
 
 @provider(IDefaultFactory)
@@ -69,6 +79,14 @@ class DateValidator(SimpleFieldValidator):
         oIds = self.context.objectIds()
         if date in oIds:
             raise Invalid(_(u'Ja existe uma agenda para esta data'))
+
+
+@indexer(IAgendaDiaria)
+def tags(obj):
+    """Indexa tags de AgendaDiaria
+    """
+    tags = obj.subjects
+    return tags
 
 
 @indexer(IAgendaDiaria)
@@ -99,8 +117,7 @@ def SearchableText_AgendaDiaria(obj):
                      if isinstance(text, basestring)])
 
 
-@indexer(IAgendaDiaria)
-def start_date(obj):
+def _start_date(obj):
     """ Converte a data da AgendaDiaria para DateTime e coloca o
         horario como 00:00:00
     """
@@ -108,6 +125,13 @@ def start_date(obj):
     # Comeco do dia
     start_date = DateTime('%s 00:00:00' % start_date.strftime('%Y-%m-%d'))
     return start_date
+
+
+@indexer(IAgendaDiaria)
+def start_date(obj):
+    """ Indexa a data de inicio com base na logica existente em _start_date
+    """
+    return _start_date(obj)
 
 
 @indexer(IAgendaDiaria)
@@ -120,3 +144,12 @@ def end_date(obj):
     end_date = DateTime('%s 23:59:59' % end_date.strftime('%Y-%m-%d'))
 
     return end_date
+
+
+@indexer(IAgendaDiaria)
+def exclude_from_nav(obj):
+    # Agendas Diarias sempre serao ocultas da navegacao
+    exclude_from_nav = obj.exclude_from_nav
+    if hasattr(exclude_from_nav, '__call__'):
+        exclude_from_nav = exclude_from_nav()
+    return exclude_from_nav